+++ CHANGES.rst
@@ -1,6 +1,29 @@
 Latest changes
 ==============
 
+In development
+--------------
+
+
+Release 1.3.2 -- 2023/08/08
+---------------------------
+
+- Fix a regression in ``joblib.Parallel`` introduced in 1.3.0 where
+  explicitly setting ``n_jobs=None`` was not interpreted as "unset".
+  https://github.com/joblib/joblib/pull/1475
+
+- Fix a regression in ``joblib.Parallel`` introduced in 1.3.0 where
+  ``joblib.Parallel`` logging methods exposed from inheritance to
+  ``joblib.Logger`` didn't work because of missing logger
+  initialization.
+  https://github.com/joblib/joblib/pull/1494
+
+- Various maintenance updates to the doc, the ci and the test.
+  https://github.com/joblib/joblib/pull/1480,
+  https://github.com/joblib/joblib/pull/1481,
+  https://github.com/joblib/joblib/pull/1476,
+  https://github.com/joblib/joblib/pull/1492
+
 Release 1.3.1 -- 2023/06/29
 ---------------------------